Allowing to execute longitudinal steps separately

+++ run.py
@@ -24,7 +24,8 @@ def run(command, env={}, ignore_errors=False):
     if process.returncode != 0 and not ignore_errors:
         raise Exception("Non zero return code: %d"%process.returncode)
 
-__version__ = open('/version').read()
+__version__ = open(os.path.join(os.path.dirname(os.path.realpath(__file__)),
+                                'version')).read()
 
 parser = argparse.ArgumentParser(description='FreeSurfer recon-all + custom template generation.')
 parser.add_argument('bids_dir', help='The directory with the input dataset '
@@ -50,6 +51,11 @@ parser.add_argument('--n_cpus', help='Number of CPUs/cores available to use.',
 parser.add_argument('--stages', help='Autorecon stages to run.',
                     choices=["autorecon1", "autorecon2", "autorecon2-cp", "autorecon2-wm", "autorecon-pial", "autorecon3", "autorecon-all", "all"],
                     default=["autorecon-all"],
+                    nargs="+")
+parser.add_argument('--steps', help='Longitudinal steps to run.',
+                    choices=[1, 2, 3],
+                    default=[1, 2, 3],
+                    type=int,
                     nargs="+")
 parser.add_argument('--template_name', help='Name for the custom group level template generated for this dataset',
                     default="average")
@@ -154,45 +160,75 @@ if args.analysis_level == "participant":
 
         timepoints = []
         if len(sessions) > 0 and longitudinal_study == True:
-            # Running each session separately, prior to doing longitudinal pipeline
-            for session_label in sessions:
-                T1s = glob(os.path.join(args.bids_dir,
-                                  "sub-%s"%subject_label,
-                                  "ses-%s"%session_label,
-                                  "anat",
-                                  "%s_T1w.nii*"%acq_tpl))
-                input_args = ""
-                for T1 in T1s:
-                    if (round(max(nibabel.load(T1).header.get_zooms()),1) < 1.0 and args.hires_mode == "auto") or args.hires_mode == "enable":
-                        input_args += " -hires"
-                    input_args += " -i %s"%T1
+            timepoints = ["sub-%s_ses-%s"%(subject_label, session_label) for session_label in sessions]
+            if (1 in args.steps):
+                # Running each session separately, prior to doing longitudinal pipeline
+                for session_label in sessions:
+                    T1s = glob(os.path.join(args.bids_dir,
+                                      "sub-%s"%subject_label,
+                                      "ses-%s"%session_label,
+                                      "anat",
+                                      "%s_T1w.nii*"%acq_tpl))
+                    input_args = ""
+                    for T1 in T1s:
+                        if (round(max(nibabel.load(T1).header.get_zooms()),1) < 1.0 and args.hires_mode == "auto") or args.hires_mode == "enable":
+                            input_args += " -hires"
+                        input_args += " -i %s"%T1
 
-                T2s = glob(os.path.join(args.bids_dir, "sub-%s"%subject_label,
-                                        "ses-%s"%session_label, "anat",
-                                        "*%s_T2w.nii*"%acq_t2))
-                FLAIRs = glob(os.path.join(args.bids_dir, "sub-%s"%subject_label,
-                                        "ses-%s"%session_label, "anat",
-                                        "*%s_FLAIR.nii*"%acq_t2))
-                if args.refine_pial == "T2":
-                    for T2 in T2s:
-                        if max(nibabel.load(T2).header.get_zooms()) < 1.2:
-                            input_args += " " + " ".join(["-T2 %s"%T2])
-                            input_args += " -T2pial"
-                elif args.refine_pial == "FLAIR":
-                    for FLAIR in FLAIRs:
-                        if max(nibabel.load(FLAIR).header.get_zooms()) < 1.2:
-                            input_args += " " + " ".join(["-FLAIR %s"%FLAIR])
-                            input_args += " -FLAIRpial"
+                    T2s = glob(os.path.join(args.bids_dir, "sub-%s"%subject_label,
+                                            "ses-%s"%session_label, "anat",
+                                            "*%s_T2w.nii*"%acq_t2))
+                    FLAIRs = glob(os.path.join(args.bids_dir, "sub-%s"%subject_label,
+                                            "ses-%s"%session_label, "anat",
+                                            "*%s_FLAIR.nii*"%acq_t2))
+                    if args.refine_pial == "T2":
+                        for T2 in T2s:
+                            if max(nibabel.load(T2).header.get_zooms()) < 1.2:
+                                input_args += " " + " ".join(["-T2 %s"%T2])
+                                input_args += " -T2pial"
+                    elif args.refine_pial == "FLAIR":
+                        for FLAIR in FLAIRs:
+                            if max(nibabel.load(FLAIR).header.get_zooms()) < 1.2:
+                                input_args += " " + " ".join(["-FLAIR %s"%FLAIR])
+                                input_args += " -FLAIRpial"
 
-                fsid = "sub-%s_ses-%s"%(subject_label, session_label)
-                timepoints.append(fsid)
-                cmd = "recon-all -subjid %s -sd %s %s -all -parallel -openmp %d"%(fsid,
-                                                                        output_dir,
-                                                                        input_args,
-                                                                        args.n_cpus)
-                resume_cmd = "recon-all -subjid %s -sd %s -all -parallel -openmp %d"%(fsid,
+                    fsid = "sub-%s_ses-%s"%(subject_label, session_label)
+                    timepoints.append(fsid)
+                    cmd = "recon-all -subjid %s -sd %s %s -all -parallel -openmp %d"%(fsid,
                                                                             output_dir,
+                                                                            input_args,
                                                                             args.n_cpus)
+                    resume_cmd = "recon-all -subjid %s -sd %s -all -parallel -openmp %d"%(fsid,
+                                                                                output_dir,
+                                                                                args.n_cpus)
+
+                    if os.path.isfile(os.path.join(output_dir, fsid,"scripts/IsRunning.lh+rh")):
+                        rmtree(os.path.join(output_dir, fsid))
+                        print("DELETING OUTPUT SUBJECT DIR AND RE-RUNNING COMMAND:")
+                        print(cmd)
+                        run(cmd)
+                    elif os.path.exists(os.path.join(output_dir, fsid)):
+                        print("SUBJECT DIR ALREADY EXISTS (without IsRunning.lh+rh), RUNNING COMMAND:")
+                        print(resume_cmd)
+                        run(resume_cmd)
+                    else:
+                        print(cmd)
+                        run(cmd)
+            
+            if (2 in args.steps):
+                # creating a subject specific template
+                input_args = " ".join(["-tp %s"%tp for tp in timepoints])
+                fsid = "sub-%s"%subject_label
+                stages = " ".join(["-" + stage for stage in args.stages])
+                cmd = "recon-all -base %s -sd %s %s %s -parallel -openmp %d"%(fsid,
+                                                                    output_dir,
+                                                                    input_args,
+                                                                    stages,
+                                                                    args.n_cpus)
+                resume_cmd = "recon-all -base %s -sd %s %s -parallel -openmp %d"%(fsid,
+                                                                        output_dir,
+                                                                        stages,
+                                                                        args.n_cpus)
 
                 if os.path.isfile(os.path.join(output_dir, fsid,"scripts/IsRunning.lh+rh")):
                     rmtree(os.path.join(output_dir, fsid))
@@ -206,49 +242,23 @@ if args.analysis_level == "participant":
                 else:
                     print(cmd)
                     run(cmd)
+            
+            if (3 in args.steps):
+                for tp in timepoints:
+                    # longitudinally process all timepoints
+                    fsid = "sub-%s"%subject_label
+                    stages = " ".join(["-" + stage for stage in args.stages])
+                    cmd = "recon-all -long %s %s -sd %s %s -parallel -openmp %d"%(tp,
+                                                                        fsid,
+                                                                        output_dir,
+                                                                        stages,
+                                                                        args.n_cpus)
 
-            # creating a subject specific template
-            input_args = " ".join(["-tp %s"%tp for tp in timepoints])
-            fsid = "sub-%s"%subject_label
-            stages = " ".join(["-" + stage for stage in args.stages])
-            cmd = "recon-all -base %s -sd %s %s %s -parallel -openmp %d"%(fsid,
-                                                                output_dir,
-                                                                input_args,
-                                                                stages,
-                                                                args.n_cpus)
-            resume_cmd = "recon-all -base %s -sd %s %s -parallel -openmp %d"%(fsid,
-                                                                    output_dir,
-                                                                    stages,
-                                                                    args.n_cpus)
-
-            if os.path.isfile(os.path.join(output_dir, fsid,"scripts/IsRunning.lh+rh")):
-                rmtree(os.path.join(output_dir, fsid))
-                print("DELETING OUTPUT SUBJECT DIR AND RE-RUNNING COMMAND:")
-                print(cmd)
-                run(cmd)
-            elif os.path.exists(os.path.join(output_dir, fsid)):
-                print("SUBJECT DIR ALREADY EXISTS (without IsRunning.lh+rh), RUNNING COMMAND:")
-                print(resume_cmd)
-                run(resume_cmd)
-            else:
-                print(cmd)
-                run(cmd)
-
-            for tp in timepoints:
-                # longitudinally process all timepoints
-                fsid = "sub-%s"%subject_label
-                stages = " ".join(["-" + stage for stage in args.stages])
-                cmd = "recon-all -long %s %s -sd %s %s -parallel -openmp %d"%(tp,
-                                                                    fsid,
-                                                                    output_dir,
-                                                                    stages,
-                                                                    args.n_cpus)
-
-                if os.path.isfile(os.path.join(output_dir, tp + ".long." + fsid,"scripts/IsRunning.lh+rh")):
-                    rmtree(os.path.join(output_dir, tp + ".long." + fsid))
-                    print("DELETING OUTPUT SUBJECT DIR AND RE-RUNNING COMMAND:")
-                print(cmd)
-                run(cmd)
+                    if os.path.isfile(os.path.join(output_dir, tp + ".long." + fsid,"scripts/IsRunning.lh+rh")):
+                        rmtree(os.path.join(output_dir, tp + ".long." + fsid))
+                        print("DELETING OUTPUT SUBJECT DIR AND RE-RUNNING COMMAND:")
+                    print(cmd)
+                    run(cmd)
 
         elif len(sessions) > 0 and longitudinal_study == False:
             # grab all T1s/T2s from multiple sessions and combine
